refactor::[routes]: relocate the email verification handler to user.ts file
